third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/show-generator-location-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that "Show Generator Location" jumps to the correct location.


Running: testIterNotStarted
Generator location revealed: [17:20]

Running: testIterSuspended1
Generator location revealed: [19:11]

Running: testIterSuspended2
Generator location revealed: [20:11]

Running: testIterSuspended3
Generator location revealed: [21:11]

Running: testIterClosed
Generator location revealed: [17:20]
